+++ irodsadmin/tests/test_commands.py
@@ -72,7 +72,7 @@ class TestIrodsOrphans(
         self.assay.save()
 
         # Create LandingZone
-        self.landing_zone = self._make_landing_zone(
+        self.landing_zone = self.make_landing_zone(
             title=ZONE_TITLE,
             project=self.project,
             user=self.as_owner.user,