+++ landingzones/tests/test_permissions.py
@@ -44,7 +44,7 @@ class TestLandingZonePermissionsBase(
         self.assay = self.study.assays.first()
 
         # Create LandingZone
-        self.landing_zone = self._make_landing_zone(
+        self.landing_zone = self.make_landing_zone(
             title=ZONE_TITLE,
             project=self.project,
             user=self.owner_as.user,