+++ landingzones/tests/test_permissions_api.py
@@ -44,7 +44,7 @@ class TestLandingZonePermissions(
         self.assay = self.study.assays.first()
 
         # Create LandingZone for project owner
-        self.landing_zone = self._make_landing_zone(
+        self.landing_zone = self.make_landing_zone(
             title=ZONE_TITLE,
             project=self.project,
             user=self.owner_as.user,