+++ landingzones/tests/test_tasks_celery.py
@@ -85,7 +85,7 @@ class TestTriggerZoneMoveTask(
         self.make_irods_colls(self.investigation)
 
         # Create zone
-        self.landing_zone = self._make_landing_zone(
+        self.landing_zone = self.make_landing_zone(
             title=ZONE_TITLE,
             project=self.project,
             user=self.user,